Create new test_hit_utils.py to test functions from utils file. Moved some tests that were in test_hit_l1a over to this new test file
